Removed misleading mention of threads in bin/nrpe-server.pl
bin/nrpe-server.pl had:

    use threads;
    <snip>
    threads->new( $daemon->start() );

On the surface this looks like this makes the daemon magically be
multithreaded and handle multiple requests in parallel.

Sadly that is not the case, as running to requests in parallel against
the $daemon will demonstrate. It is single-threaded.

$daemon->start() never returns, so threads->new() is never called. The
threads code does absolutely nothing.

So it is no-op and misleading. In addition, perldoc threads says:

    The "interpreter-based threads" provided by Perl are not the fast,
    lightweight system for multitasking that one might expect or hope for.
    Threads are implemented in a way that make them easy to misuse. Few people
    know how to use them correctly or will be able to provide help.

    The use of interpreter-based threads in perl is officially discouraged.

So I simply removed the misleading no-op threads code and replaced:

    use threads;
    <snip>
    threads->new( $daemon->start() );

with:
    $daemon->start();