Daemon fixes (#18)
* Remove syntax error from non-used documentation code

* Nagios::NRPE::Daemon: Ran perltidy on example code

* Nagios::NRPE::Daemon: Doc example - actually start the daemon that gets created

* Nagios::NRPE::Packet: Set result_code to 2324 only if it is not defined

Before result_code got set if it was false. That had the problem that
if the result_code explicity got set to STATUS_OK (which is 0 and
therefore perl false) it got set the 2324 which is a nagios error.

* Nagios::NRPE::Daemon: Fix multiple problems with callbacks

This fixes both the example callbacks and the callback in
bin/nrpe-server.pl

* Allow callback to specify status code
  Allow specifying the status code (ok, warning, critical, etc.) as
  well as output message. Before the callback could only specify the
  output message, but nagios uses both
* The substitution of $ARGN$ strings simply didn't work
* The entire STDERR+STDOUT output from the child would be sent back as
  the message. This is inconsistent with Nagios that only considers
  STDOUT and ignores STDERR
* If the daemon is called with a $check that wasn't configured, it
  wouldn't return anything and check_nrpe would say:
    CHECK_NRPE: Received 0 bytes from daemon.  Check the remote server logs for error messages.
    exit code: 3
  Now it says:
    sprintf("No such check: '%s'", $check_name)
    exit code: 3

nrpe-server.pl and the examples still don't handle exit codes more
elaborately than checking for $success, but that is due to a
limitation of IPC::Cmd::run, and could be fixed by using IPC::Run, but
that introduces a new dependency and was more hassle than I was
prepared to deal with.

* Remove misleading shebang from .pm files and leading =pod-s too

* Removed misleading mention of threads in bin/nrpe-server.pl

bin/nrpe-server.pl had:

    use threads;
    <snip>
    threads->new( $daemon->start() );

On the surface this looks like this makes the daemon magically be
multithreaded and handle multiple requests in parallel.

Sadly that is not the case, as running to requests in parallel against
the $daemon will demonstrate. It is single-threaded.

$daemon->start() never returns, so threads->new() is never called. The
threads code does absolutely nothing.

So it is no-op and misleading. In addition, perldoc threads says:

    The "interpreter-based threads" provided by Perl are not the fast,
    lightweight system for multitasking that one might expect or hope for.
    Threads are implemented in a way that make them easy to misuse. Few people
    know how to use them correctly or will be able to provide help.

    The use of interpreter-based threads in perl is officially discouraged.

So I simply removed the misleading no-op threads code and replaced:

    use threads;
    <snip>
    threads->new( $daemon->start() );

with:
    $daemon->start();